requirements.txt
requests>=2.0
fonttools>=4.0
Pillow
